refactor(myrecipes): extract TweakEditor to its own file (one component per file)
